feature: - added factory to create "Problems" (ProblemsFactory) refactor: - changed data storage to parameterized types (instead of .bin file)
