--- src/main/java/akademia/assistant/front/genericRepo/Product2.java
@@ -1,4 +1,0 @@
-package akademia.assistant.front.genericRepo;
-
-public class Product2 {
-}
