--- src/main/java/akademia/assistant/front/genericRepo/Repo.java
@@ -1,11 +1,0 @@
-package akademia.assistant.front.genericRepo;
-
-import java.util.List;
-
-public interface Repo<T> {
-
-    void add(T element);
-    List<T> getElements();
-
-    void test();
-}
