+++ src/main/java/akademia/assistant/front/model/User.java
@@ -1,8 +1,6 @@
 package akademia.assistant.front.model;
 
-import java.io.Serializable;
-
-public class User implements Serializable {
+public class User {
     private String username = "";
     private String password = "";
