--- src/main/java/akademia/assistant/front/repository/BinFileRepository.java
@@ -1,61 +1,0 @@
-package akademia.assistant.front.repository;
-
-import akademia.assistant.front.model.Category;
-import akademia.assistant.front.model.User;
-
-import java.io.*;
-import java.util.ArrayList;
-import java.util.List;
-
-public class BinFileRepository implements Repository {
-    @Override
-    public String saveUsers(List<User> users) {
-        try {
-            ObjectOutputStream oos = new ObjectOutputStream(new FileOutputStream("users.bin"));
-            oos.writeObject(users);
-            oos.flush();
-            oos.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return "Users have been saved.";
-    }
-
-    @Override
-    public List<User> loadUsers() {
-        try {
-            ObjectInputStream ois = new ObjectInputStream(new FileInputStream("users.bin"));
-            List<User> users = (List<User>) ois.readObject();
-            ois.close();
-            return users;
-        } catch (IOException | ClassNotFoundException e) {
-            return new ArrayList<>();
-        }
-    }
-
-    @Override
-    public String updateCategoriesToFile(List<Category> categories) {
-        try {
-            new File("categories.bin").delete();
-            ObjectOutputStream oos = new ObjectOutputStream(new FileOutputStream("categories.bin"));
-            oos.writeObject(categories);
-            oos.flush();
-            oos.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return "Users have been saved.";
-    }
-
-    @Override
-    public List<Category> loadCategoriesFromFile() {
-        try {
-            ObjectInputStream ois = new ObjectInputStream(new FileInputStream("categories.bin"));
-            List<Category> categories = (List<Category>) ois.readObject();
-            ois.close();
-            return categories;
-        } catch (IOException | ClassNotFoundException e) {
-            return new ArrayList<>();
-        }
-    }
-}
