--- src/main/java/akademia/assistant/front/repository/CategoryRepository.java
@@ -1,5 +1,0 @@
-package akademia.assistant.front.repository;
-
-public class CategoryRepository {
-
-}
