--- src/main/java/akademia/assistant/front/repository/ProblemRepository.java
@@ -1,4 +1,0 @@
-package akademia.assistant.front.repository;
-
-public class ProblemRepository {
-}
