+++ src/main/java/akademia/assistant/front/repository/Repository.java
@@ -1,16 +1,10 @@
 package akademia.assistant.front.repository;
 
-import akademia.assistant.front.model.User;
-import akademia.assistant.front.model.Category;
-
 import java.util.List;
 
-public interface Repository {
-    String saveUsers(List<User> users);
+public interface Repository<T> {
 
-    List<User> loadUsers();
+    void addElement(T element);
 
-    String updateCategoriesToFile(List<Category> categories);
-
-    List<Category> loadCategoriesFromFile();
+    List<T> getElements();
 }
