+++ src/main/java/akademia/assistant/front/service/Service.java
@@ -1,28 +1,25 @@
 package akademia.assistant.front.service;
 
+import akademia.assistant.front.controller.factory.ProblemsFactory;
 import akademia.assistant.front.model.User;
 import akademia.assistant.front.exception.AuthenticationException;
 import akademia.assistant.front.model.Category;
-import akademia.assistant.front.model.Problem;
-import akademia.assistant.front.repository.BinFileRepository;
 import akademia.assistant.front.repository.Repository;
+import akademia.assistant.front.repository.RepositoryImpl;
 
-import java.util.ArrayList;
 import java.util.List;
 //todo refactor to two main services: for Authentication and for basic operations on Problem
 public class Service {
-    private final Repository repository = new BinFileRepository();
-    private final List<Category> categoryList;
-    private final List<User> users;
+    private final Repository<User> usersRepository = new RepositoryImpl<>();
+    private final Repository<Category> categoryRepository = new RepositoryImpl<>();
     private User currentUser;
 
     public Service() {
-        users = repository.loadUsers();
-//        categoryList = repository.loadCategoriesFromFile();
-        categoryList = new ArrayList<>();
-        categoryList.add(new Category("Basic", basicProblemFactory()));
-        categoryList.add(new Category("OOP", OOPProblemFactory()));
-        categoryList.add(new Category("Advance", advanceProblemFactory()));
+        usersRepository.addElement(new User("qwe", "qwe"));//added to inspect login possibility
+        ProblemsFactory problemFactory = new ProblemsFactory();
+        categoryRepository.addElement(new Category("Basic", problemFactory.basicProblems()));
+        categoryRepository.addElement(new Category("OOP", problemFactory.OOPProblems()));
+        categoryRepository.addElement(new Category("Advance", problemFactory.advanceProblems()));
     }
 
     public void login(String username, String password) {
@@ -33,7 +30,7 @@ public class Service {
     }
 
     private boolean doesUserExist(String username, String password) {
-        for (User user : users) {
+        for (User user : usersRepository.getElements()) {
             if (user.getUsername().equals(username) && user.getPassword().equals(password)) {
                 return true;
             }
@@ -54,7 +51,7 @@ public class Service {
 
 
     private boolean doesNicknameExist(String newUsername) {
-        for (User user : users) {
+        for (User user : usersRepository.getElements()) {
             if (user.getUsername().equals(newUsername)) {
                 return true;
             }
@@ -63,8 +60,7 @@ public class Service {
     }
 
     private void addUserToUserList(User newUser) {
-        users.add(newUser);
-        repository.saveUsers(users);
+        usersRepository.addElement(newUser);
     }
 
     private void clearCurrentUser() {
@@ -75,54 +71,19 @@ public class Service {
         return currentUser;
     }
 
-    private void updateCategories() {
+    /*private void updateCategories() {
         repository.updateCategoriesToFile(categoryList);
-    }
+    }*/
 
     public List<Category> getCategories() {
-        return categoryList;
+        return categoryRepository.getElements();
     }
 
     public int getAmountCategories() {
-        return categoryList.size();
+        return categoryRepository.getElements().size();
     }
 
     public Category getCategoryByIndex(int index) {
-        return categoryList.get(index);
-    }
-
-    public List<Problem> basicProblemFactory() {
-        Problem basics1 = new Problem("Wyświetlanie w konsoli",
-                "Jak wygląda instrukcja, która drukuje w języku Java?");
-        Problem basics2 = new Problem("Tytuł problemu 2", "Pytanie problemu 2");
-        Problem basics3 = new Problem("Tytuł problemu 3", "Pytanie problemu 3");
-        List<Problem> basicProblems = new ArrayList<>();
-        basicProblems.add(basics1);
-        basicProblems.add(basics2);
-        basicProblems.add(basics3);
-        return basicProblems;
-    }
-
-    public List<Problem> OOPProblemFactory() {
-        Problem OOP1 = new Problem("Tworzenie obiektu",
-                "Jaką instrukcją należy stworzyć obiekt przykładowej klasy Controller?");
-        Problem OOP5 = new Problem("Tytuł problemu 5", "Pytanie problemu 5");
-        Problem OOP6 = new Problem("Tytuł problemu 6", "Pytanie problemu 6");
-        List<Problem> OOPProblems = new ArrayList<>();
-        OOPProblems.add(OOP1);
-        OOPProblems.add(OOP5);
-        OOPProblems.add(OOP6);
-        return OOPProblems;
-    }
-
-    public List<Problem> advanceProblemFactory() {
-        Problem advance1 = new Problem("Tytuł problemu 10",
-                "Pytanie problemu 10?");
-        Problem advance2 = new Problem("Tytuł problemu 11",
-                "Pytanie problemu 11?");
-        List<Problem> OOPProblems = new ArrayList<>();
-        OOPProblems.add(advance1);
-        OOPProblems.add(advance2);
-        return OOPProblems;
+        return categoryRepository.getElements().get(index);
     }
 }
